openings-collection/requirements.txt
flask==2.2.2
python-dotenv==0.21.0
pymongo==3.12.0
Flask-PyMongo==2.3.0
pytest==7.2.0